configury: use "project dir" instead of "project name"
This change basically renames a project based on its directory name.
This minimizes the changes of pmix_mca.m4 compared to its upstream
Open MPI opal_mca.m4 counterpart, but is a little weird (such that a
project is really known by its directory name, not the name that
people want to call it).

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>