Simplify the PMIx_get processing:
If the key starts with “pmix”, then it is a PMIx-namespace key and can only be in the data provided at startup (which is stored in the “internal" hash table). Otherwise, it is data that was PMix_Put and is in the “modex” hash table. PMIx job-level data is stored in the “internal” hash table against rank=PMIX_RANK_WILDCARD.

This reduces all calls to PMIx_Get to a single hash-table lookup

Fix the problem of a proc asking for job-level data from another nspace by ensuring that the server detects the situation, and then returns the nspace blob for the other nspace.

Backport of Coverity warning fixes

Address review line comments from @artpol84.

Backport changes from OMPI