Added the register nspace invoke for all active GDS modules.
This fix is required when there are more than one active GDS modules.
Server must prepare a nspace for each active GDS module
before client is started. Because server doesn't know what kind of
module will be used on the client side when it start. Consequently,
there must be a nspace deregistration for each GDS during the finilize.

Signed-off-by: Boris Karasev <karasev.b@gmail.com>